+++ src/drivers/uart.h
@@ -3,4 +3,4 @@
 void uart_init(unsigned baud_rate);
 void uart_write_char(char c);
 void uart_write_string(const char *s);
-void uart_write_buf(const char *s, size_t len);
+void uart_write_buf(const char *s, int len);
