+++ src/lib/include/unistd.h
@@ -2,5 +2,5 @@
 
 #include "../../drivers/uart.h"
 
-int write(int fd, const char *buf, size_t len);
-int read(int fd, char *buf, size_t len);
+int write(int fd, const char *buf, int len);
+int read(int fd, char *buf, int len);
