+++ src/lib/unistd.c
@@ -1,11 +1,11 @@
-#include "lib/unistd.h"
+#include "include/unistd.h"
 
-int write(int fd, const char *buf, size_t len) {
+int write(int fd, const char *buf, int len) {
     if (fd != 0) return -1;
     uart_write_buf(buf, len);
 }
 
-int read(int fd, char *buf, size_t len) {
+int read(int fd, char *buf, int len) {
     // TODO: Write something
     return -1;
 }
